feat: polish signal routes
