+++ apps/torus-portal/src/app/(pages)/signals/create-signal/_components/create-signal-slider-field.tsx
